# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API E-commerce Fashion",
        "description": "Hợp đồng API giữa Vue.js Frontend và Laravel Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/coupons": {
            "get": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Lấy danh sách mã giảm giá (có phân trang & lọc)",
                "operationId": "9a3c46a3dccfbfeda99db62d75269908",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm theo mã code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Lọc theo loại giảm giá",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "percent",
                                "fixed"
                            ]
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Lọc theo trạng thái (1=active, 0=inactive)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số bản ghi mỗi trang (mặc định: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Tạo mã giảm giá mới",
                "operationId": "19d36675c30b185c7085a17ff51de9a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "type",
                                    "discount_value",
                                    "is_active"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Mã coupon, unique",
                                        "type": "string",
                                        "example": "SUMMER20"
                                    },
                                    "type": {
                                        "description": "Loại giảm giá",
                                        "type": "string",
                                        "example": "percent",
                                        "enum": [
                                            "percent",
                                            "fixed"
                                        ]
                                    },
                                    "discount_value": {
                                        "description": "Giá trị giảm (% hoặc VNĐ)",
                                        "type": "number",
                                        "example": 20
                                    },
                                    "price_min_order_value": {
                                        "description": "Đơn hàng tối thiểu để áp dụng",
                                        "type": "number",
                                        "example": 500000,
                                        "nullable": true
                                    },
                                    "max_usage": {
                                        "description": "Số lần sử dụng tối đa",
                                        "type": "integer",
                                        "example": 100,
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "expiry_date": {
                                        "description": "Ngày hết hạn (phải sau hôm nay)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mã giảm giá đã được thêm thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lỗi validate dữ liệu"
                    }
                }
            }
        },
...
        "/api/admin/coupons/{coupon}": {
            "get": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Xem chi tiết một mã giảm giá",
                "operationId": "5b465292bf8c044a258ebef0e8489104",
                "parameters": [
                    {
                        "name": "coupon",
                        "in": "path",
                        "description": "ID của mã giảm giá",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy mã giảm giá"
                    }
                }
            },
            "put": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Cập nhật mã giảm giá",
                "operationId": "c7f9a942933a3536efe7008f2cc3dfe3",
                "parameters": [
                    {
                        "name": "coupon",
                        "in": "path",
                        "description": "ID của mã giảm giá cần cập nhật",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "type",
                                    "discount_value",
                                    "is_active"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Mã coupon (unique, bỏ qua coupon hiện tại)",
                                        "type": "string",
                                        "example": "SUMMER20"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "fixed",
                                        "enum": [
                                            "percent",
                                            "fixed"
                                        ]
                                    },
                                    "discount_value": {
                                        "type": "number",
                                        "example": 50000
                                    },
                                    "price_min_order_value": {
                                        "type": "number",
                                        "example": 300000,
                                        "nullable": true
                                    },
                                    "max_usage": {
                                        "type": "integer",
                                        "example": 50,
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "expiry_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mã giảm giá đã được cập nhật thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "422": {
                        "description": "Lỗi validate dữ liệu"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Xóa mã giảm giá",
                "operationId": "859d8c15782295822a57646982d77fa1",
                "parameters": [
                    {
                        "name": "coupon",
                        "in": "path",
                        "description": "ID của mã giảm giá cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mã giảm giá đã được xóa thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
        "/api/admin/coupons/check": {
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Kiểm tra mã giảm giá có hợp lệ hay không",
                "operationId": "3ddb3ae9a1ec6eb47673c4e20417ae8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "order_total"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "SUMMER20"
                                    },
                                    "order_total": {
                                        "type": "number",
                                        "example": 500000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trả về trạng thái hợp lệ hoặc không hợp lệ của coupon"
                    }
                }
            }
        },
...
    },
...
}